requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
psycopg2-binary>=2.9.9
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.6.1
email-validator>=2.0.0
python-dotenv==1.0.1
reportlab==4.2.5
PyPDF2==3.0.1
openai>=1.54.0
firebase-admin==6.5.0
requests>=2.31.0
Pillow>=10.0.0

